[FRONTEND] Extract AccountsTableHeader from AccountList
